Fixed bug that caused wrong Jacobian on multiple processors.
Needed to update the voltage variables p_VD and p_VQ from the values stored in the
buffer pointed by p_VDQptr. This pointer is set in the setXCBuf routine and it
points to location where GridPACK stores the values being exchanged. When
updateBuses() is called, this buffer gets updated. So, any functions called
after updateBuses() needed to retrieve the updated ghost values from this
buffer.